+++ src/Interfaces/WordPress/Roles/ModifyUserRoleInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace ObsidianMoon\ProjectUtilities\Interfaces\WordPress\Roles;
+
+interface ModifyUserRoleInterface
+{
+    public function addRole(): void;
+}